+++ src/types/jsx.ts
@@ -24,6 +24,8 @@ declare global {
     'u-widget': import('../elements/u-widget.js').UWidget;
   }
 
+  // JSX.IntrinsicElements 전역 증강은 namespace 선언만 가능 (ES module 문법 대체 불가)
+  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace JSX {
     interface IntrinsicElements {
       'u-widget': UWidgetElementProps;